Send notification when video meeting is started.
