+++ server/socket/videoChat.ts
@@ -1,5 +1,7 @@
 import { Socket, Server } from "socket.io";
 import mediasoup, { types as mediasoupTypes } from "mediasoup";
+import { sendingMessages } from "../models/messages.js";
+import { Types } from "mongoose";
 
 const HOST_IP = process.env.HOST_IP || "127.0.0.1";
 const rtcMinPort: number = Number(process.env.MEDIASOUP_RTCMINPORT) ?? 2000;
@@ -9,7 +11,9 @@ const videoChat = function () {
   // @ts-ignore
   const _self = this;
 
-  _self.init = async function (connections: Server) {
+  _self.init = async function (io: Server) {
+    const videoConnection = io.of("/mediasoup");
+
     const createWorker = async () => {
       const worker = await mediasoup.createWorker({
         rtcMinPort,
@@ -106,7 +110,7 @@ const videoChat = function () {
       },
     ];
 
-    connections.on("connection", async (socket: Socket) => {
+    videoConnection.on("connection", async (socket: Socket) => {
       console.log("Video Chat connection", socket.id);
       socket.emit("connection-success", {
         socketId: socket.id,
@@ -137,62 +141,77 @@ const videoChat = function () {
 
           const { roomName } = peers[socket.id];
           delete peers[socket.id];
+          delete rooms[roomName];
 
           // remove socket from room
-          rooms[roomName] = {
-            router: rooms[roomName].router,
-            peers: rooms[roomName].peers?.filter((socketId) => socketId !== socket.id),
-          };
+          // rooms[roomName] = {
+          //   router: rooms[roomName].router,
+          //   peers: rooms[roomName].peers?.filter((socketId) => socketId !== socket.id),
+          // };
         } catch (err) {
           console.log("Disconnect Error.");
           console.log(err);
         }
       });
 
-      socket.on("joinRoom", async ({ roomName }: { roomName: string }, callback: Function) => {
-        console.log(roomName);
-        const createRoom = async (socketId: string) => {
-          let router1;
-          let currentPeers: string[] = [];
-          if (rooms[roomName]) {
-            router1 = rooms[roomName].router;
-            currentPeers = rooms[roomName].peers || [];
-          } else {
-            router1 = await worker.createRouter({ mediaCodecs });
-          }
+      socket.on(
+        "joinRoom",
+        async (
+          { roomName, workspace, token }: { roomName: string; workspace: Types.ObjectId; token: string },
+          callback: Function
+        ) => {
+          console.log(roomName);
+          const createRoom = async (socketId: string) => {
+            let router1;
+            let currentPeers: string[] = [];
+            if (rooms[roomName]) {
+              router1 = rooms[roomName].router;
+              currentPeers = rooms[roomName].peers || [];
+            } else {
+              router1 = await worker.createRouter({ mediaCodecs });
+              const message = "I've launch a video meeting, click the video button to join in!";
+              await sendingMessages(
+                io,
+                token,
+                { workspace, type: "team", id: roomName as unknown as Types.ObjectId },
+                message,
+                undefined
+              );
+            }
 
-          console.log(`Router ID: ${router1.id}`, currentPeers.length);
+            console.log(`Router ID: ${router1.id}`, currentPeers.length);
 
-          rooms[roomName] = {
-            router: router1,
-            peers: [...currentPeers, socketId],
+            rooms[roomName] = {
+              router: router1,
+              peers: [...currentPeers, socketId],
+            };
+
+            return router1;
           };
 
-          return router1;
-        };
+          // create Router if it does not exist
+          // const router1 = rooms[roomName] && rooms[roomName].get('data').router || await createRoom(roomName, socket.id)
+          const router1 = await createRoom(socket.id);
 
-        // create Router if it does not exist
-        // const router1 = rooms[roomName] && rooms[roomName].get('data').router || await createRoom(roomName, socket.id)
-        const router1 = await createRoom(socket.id);
+          peers[socket.id] = {
+            socket,
+            roomName, // Name for the Router this Peer joined
+            transports: [],
+            producers: [],
+            consumers: [],
+            peerDetails: {
+              name: "",
+              isAdmin: false, // Is this Peer the Admin?
+            },
+          };
 
-        peers[socket.id] = {
-          socket,
-          roomName, // Name for the Router this Peer joined
-          transports: [],
-          producers: [],
-          consumers: [],
-          peerDetails: {
-            name: "",
-            isAdmin: false, // Is this Peer the Admin?
-          },
-        };
+          // get Router RTP Capabilities
+          const { rtpCapabilities } = router1;
 
-        // get Router RTP Capabilities
-        const { rtpCapabilities } = router1;
-
-        // call callback from the client and send back the rtpCapabilities
-        callback({ rtpCapabilities });
-      });
+          // call callback from the client and send back the rtpCapabilities
+          callback({ rtpCapabilities });
+        }
+      );
 
       const createWebRtcTransport = (mediasoupRouter: mediasoupTypes.Router): Promise<mediasoupTypes.WebRtcTransport> =>
         new Promise((resolve, reject) => {
